refactoring pymilo's Import/Export exceptions
